[astro] Handle istio-proxy for Kubernetes Pods (#62)
Istio service mesh is not compatible by default with Kubernetes Jobs.
The normal behavior is that a Job will be started, get an istio-proxy
sidecar attached to it via the istio mutating webhook, run until
completion, then the 'main' container in the pod stops, but istio-proxy
hangs around indefinitely. This change handles cleanly exiting the
Istio sidecar 'istio-proxy' when a Kubernetes Executor task completes.

Initially committed in 84fa48f. This ports it to Provider which
is neccessary for KubernetesPodOperator users.